Edit to spec to clarify scope
Building on @jyasskin's amendments, I propose a compromise to clarify that GPC is intended to allow users to opt out of sale, sharing, or the use of their data for cross-context targeting. However, I note that the laws that have been enacted to date generally only restrict the use of cross-organization data for targeting.